Importation des fichiers du prof pour repartir avec une bonne base

+++ score-valider.php
@@ -30,11 +30,12 @@ if(strlen($_GET['id']))
 		header("Location: ".SITE_HTTP."admin.php?message=$message");
 		exit;
 	}
+    
 ?><!DOCTYPE html>
 <html lang="fr">
 	<head>
 		<meta charset="UTF-8">
-		<title>Valider un score | Guitar Wars</title>
+		<title>Valider un score | <?php echo SITE_NOM; ?></title>
 		<link rel="stylesheet" type="text/css" href="style.css" />
 	</head>
 	<body>
@@ -47,7 +48,7 @@ if(strlen($_GET['id']))
         </p>
         <?php    
             //Si un screenshot existe --> on l'affiche
-            if(!empty($score['screenshot']))
+             if(!empty($score['screenshot']) and is_file(UPLOAD_PHOTOS.$score['screenshot']))
                 echo '<p><img src="'.SITE_IMAGES.$score['screenshot'].'" alt="screenshot"/></p>';
         ?>    
             <form method="post" action="<?php echo $_SERVER['PHP_SELF']; ?>">
